+++ test/ApiTest.php
@@ -1,15 +1,16 @@
 <?php
+
 /**
- * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)
- * @copyright Copyright (c) 2015 Apiary Ltd. <support@apiary.io>
+ * @see       https://github.com/laminas-api-tools/api-tools-documentation-apiblueprint for the canonical source repository
+ * @copyright https://github.com/laminas-api-tools/api-tools-documentation-apiblueprint/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas-api-tools/api-tools-documentation-apiblueprint/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZFTest\Apigility\Documentation\ApiBlueprint;
+namespace LaminasTest\ApiTools\Documentation\ApiBlueprint;
 
+use Laminas\ApiTools\Documentation\Api as BaseApi;
+use Laminas\ApiTools\Documentation\ApiBlueprint\Api;
 use PHPUnit_Framework_TestCase as TestCase;
-use ZF\Apigility\Documentation\Api as BaseApi;
-use ZF\Apigility\Documentation\ApiBlueprint\Api;
 
 class ApiTest extends TestCase
 {